Fixed race condition:
Events are placed on a queue. Events include:
 - Incoming data (lastMessage updated)
 - Updating of data in useState.

Hence, when two messages come in from the server 'quickly', and one of them is a Devices list,
then what happens is this:
 - The Devices list is processed (calling setDevices(...))
 - Then the next message is processed.
 - Then the devices list (stored in a state) is actually updated.

Hence if the processing of 2nd message is looking up data in the devices list, then nothing
is found because it is not updated yet.

The solution: To use a useRef construction. When the devices message is processed, both calling setDeivces
and the ref is updated.

The ref does not trigger a state change and thus a re-rendering.